another small step

+++ wowViewerLib/src/glsl/vulkan/m2Shader.vert
@@ -9,7 +9,7 @@
 /* vertex shader code */
 layout(location=0) in vec3 aPosition;
 layout(location=1) in vec3 aNormal;
-layout(location=2) in vec4 bones;
+layout(location=2) in uvec4 bones;
 layout(location=3) in vec4 boneWeights;
 layout(location=4) in vec2 aTexCoord;
 layout(location=5) in vec2 aTexCoord2;
